backend/requirements.txt
# AnasFlow Backend Requirements
# Install: pip install -r requirements.txt

# Web Framework
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9

# Database (Supabase uses PostgreSQL)
supabase>=2.3.4
psycopg2-binary
sqlalchemy>=2.0.27

# Authentication
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
google-auth>=2.28.1

# HTTP Requests (for API integrations)
httpx>=0.26.0
requests>=2.31.0
aiohttp>=3.9.3

# Environment Variables
python-dotenv>=1.0.1

# Validation
pydantic>=2.6.1
pydantic-settings>=2.1.0
email-validator>=2.1.0

# CORS
fastapi-cors>=0.0.6

# Date/Time
python-dateutil>=2.8.2

# Async Support
asyncio>=3.4.3

# Stripe (for payments)
stripe>=8.2.0

# Monitoring
sentry-sdk>=1.40.0

# AI APIs
openai>=1.12.0

# Rate Limiting
slowapi>=0.1.9
